Ajout d’un sémaphore indiquant le nombre de places restante dans le serveur

+++ Sprint_2/serveur/global.c
@@ -2,4 +2,4 @@
 
 Client tabClient[MAX_CLIENT];
 pthread_t tabThread[MAX_CLIENT];
-long nbConnectedClient = 0;
+int nbConnectedClient = 0;
